+++ packages/gestalt-codemods/0.111.0-0.112.0/__testfixtures__/heading-remove-semibold-dynamic.input.js
@@ -1,0 +1,11 @@
+// @flow
+import React from 'react';
+import { Heading } from 'gestalt';
+
+export default function BoldHeading() {
+  return (
+    <Heading size="xl" weight={"bold".length > 10 ? 'bold' : 'semibold'}>
+      Bold
+    </Heading>
+  );
+}