+++ packages/gestalt-codemods/0.111.0-0.112.0/__testfixtures__/text-remove-semibold-static.output.js
@@ -1,0 +1,11 @@
+// @flow
+import React from 'react';
+import { Text } from 'gestalt';
+
+export default function BoldText() {
+  return (
+    <Text size="xl" weight="bold">
+      Bold
+    </Text>
+  );
+}